Changed from Integer to Double

+++ app/src/main/java/com/example/tyler/calculatornew/AddInteger.java
@@ -1,18 +1,18 @@
 package com.example.tyler.calculatornew;
 
-public class AddInteger extends BinaryOperator<Integer> {
+public class AddInteger extends BinaryOperator<Double> {
 
 
     @Override
-    protected Integer calc(Integer a, Integer b) {
+    protected Double calc(Double a, Double b) {
         double total = a+b;
 
-        if (total > Integer.MAX_VALUE) {
+        if (total > Double.MAX_VALUE) {
             throw new ArithmeticException("Integer value overflow");
         }
-        if (total < Integer.MIN_VALUE) {
+        if (total < Double.MIN_VALUE) {
             throw new ArithmeticException("Integer value underflow");
         }
-        return (int) total;
+        return  total;
     }
 }
